feat(rebuild): SignatureSet, SignatureSetArray and ver_mult_agg_sig
