+++ rebuild/src/addon.cc
@@ -189,184 +189,6 @@ Uint8ArrayArgArray::Uint8ArrayArgArray(
         }
     }
 };
-/**
- *
- *
- * TestWorker
- *
- *
- */
-class TestWorker : public BlstAsyncWorker
-{
-public:
-    enum TestSyncOrAsync
-    {
-        SYNC = 0,
-        ASYNC = 1,
-    };
-    enum TestPhase
-    {
-        SETUP = 0,
-        EXECUTION = 1,
-        RETURN = 2
-    };
-    enum TestCase
-    {
-        NORMAL_EXECUTION = -1,
-        SET_ERROR = 0,
-        THROW_ERROR = 1,
-        UINT_8_ARRAY_ARG = 2,
-        UINT_8_ARRAY_ARG_ARRAY = 3,
-        PUBLIC_KEY_ARG = 4,
-        PUBLIC_KEY_ARG_ARRAY = 5,
-        SIGNATURE_ARG = 6,
-        SIGNATURE_ARG_ARRAY = 7
-    };
-
-public:
-    TestWorker(const Napi::CallbackInfo &info)
-        : BlstAsyncWorker{info},
-          _test_phase{TestPhase::SETUP},
-          _test_case{0},
-          _return_value{} {}
-
-protected:
-    void
-    Setup() override
-    {
-        Napi::Value test_phase_value = _info[1];
-        if (!test_phase_value.IsNumber())
-        {
-            SetError("testPhase must be a TestPhase enum");
-            return;
-        }
-        _test_phase = static_cast<TestPhase>(test_phase_value.As<Napi::Number>().Uint32Value());
-        Napi::Value test_case_value = _info[2];
-        if (!test_case_value.IsNumber())
-        {
-            SetError("testCase must be a TestCase enum");
-            return;
-        }
-        _test_case = test_case_value.As<Napi::Number>().Int32Value();
-        if (_test_phase == TestPhase::SETUP)
-        {
-            switch (_test_case)
-            {
-            case TestCase::SET_ERROR:
-                SetError("setup: TestCase.SET_ERROR");
-                break;
-            case TestCase::THROW_ERROR:
-                throw Napi::Error::New(_env, "setup: TestCase.THROW_ERROR");
-                break;
-            case TestCase::UINT_8_ARRAY_ARG:
-            {
-                Uint8ArrayArg a{_env, _info[3], "TEST"};
-                if (a.HasError())
-                {
-                    SetError(a.GetError());
-                    return;
-                }
-                break;
-            }
-            case TestCase::UINT_8_ARRAY_ARG_ARRAY:
-            {
-                Uint8ArrayArgArray a{_env, _info[3], "TEST", "TESTS"};
-                if (a.HasError())
-                {
-                    SetError(a.GetError());
-                    return;
-                }
-                break;
-            }
-            case TestCase::PUBLIC_KEY_ARG:
-            {
-                PublicKeyArg a{_env, _info[3]};
-                if (a.HasError())
-                {
-                    SetError(a.GetError());
-                    return;
-                }
-                break;
-            }
-            case TestCase::PUBLIC_KEY_ARG_ARRAY:
-            {
-                PublicKeyArgArray a{_env, _info[3]};
-                if (a.HasError())
-                {
-                    SetError(a.GetError());
-                    return;
-                }
-                break;
-            }
-            case TestCase::SIGNATURE_ARG:
-            {
-                SignatureArg a{_env, _info[3]};
-                if (a.HasError())
-                {
-                    SetError(a.GetError());
-                    return;
-                }
-                break;
-            }
-            case TestCase::SIGNATURE_ARG_ARRAY:
-            {
-                SignatureArgArray a{_env, _info[3]};
-                if (a.HasError())
-                {
-                    SetError(a.GetError());
-                    return;
-                }
-                break;
-            }
-            }
-        }
-    }
-    void Execute() override
-    {
-        if (_test_phase != TestPhase::EXECUTION)
-        {
-            _return_value.append("VALID_TEST");
-            return;
-        }
-
-        switch (_test_case)
-        {
-        case TestCase::SET_ERROR:
-            SetError("execution: TestCase.SET_ERROR");
-            break;
-        case TestCase::THROW_ERROR:
-            throw std::exception();
-            break;
-        case -1:
-        default:
-            _return_value.append("VALID_TEST");
-        }
-    }
-    Napi::Value GetReturnValue() override
-    {
-        if (_test_phase != TestPhase::RETURN)
-        {
-            return Napi::String::New(_env, _return_value);
-        }
-        switch (_test_case)
-        {
-        case TestCase::SET_ERROR:
-            SetError("return: TestCase.SET_ERROR");
-            break;
-        case TestCase::THROW_ERROR:
-            throw Napi::Error::New(_env, "return: TestCase.THROW_ERROR");
-            break;
-        default:
-            SetError("return: unknown test case");
-        }
-        return _env.Undefined();
-    }
-
-private:
-    TestPhase _test_phase;
-    int32_t _test_case;
-    std::string _return_value;
-};
 
 /**
  *
@@ -396,7 +218,6 @@ BlstTsAddon::BlstTsAddon(Napi::Env env, Napi::Object exports)
 {
     DefineAddon(exports, {
                              InstanceValue("BLST_CONSTANTS", BuildJsConstants(env), napi_enumerable),
-                             InstanceMethod("runTest", &BlstTsAddon::RunTest, napi_enumerable),
                          });
     SecretKey::Init(env, exports, this);
     PublicKey::Init(env, exports, this);
@@ -426,21 +247,6 @@ bool BlstTsAddon::GetRandomBytes(blst::byte *ikm, size_t length)
     } while (1 == RAND_poll());
 
     return false;
-}
-Napi::Value BlstTsAddon::RunTest(const Napi::CallbackInfo &info)
-{
-    if (!info[0].IsNumber())
-    {
-        throw Napi::TypeError::New(info.Env(), "First argument must be enum TestSyncOrAsync");
-    }
-    int32_t sync_or_async = info[0].ToNumber().Int32Value();
-    if (sync_or_async == 0)
-    {
-        TestWorker worker{info};
-        return worker.RunSync();
-    }
-    TestWorker *worker = new TestWorker{info};
-    return worker->Run();
 }
 Napi::Object BlstTsAddon::BuildJsConstants(Napi::Env &env)
 {
