+++ rebuild/src/functions.h
@@ -10,11 +10,46 @@
 #include "public_key.h"
 #include "signature.h"
 
+namespace
+{
+    class SignatureSet : public BlstBase
+    {
+    public:
+        Uint8ArrayArg _msg;
+        PublicKeyArg _publicKey;
+        SignatureArg _signature;
+
+        SignatureSet(Napi::Env env, const Napi::Value &raw_arg);
+
+        // non-copyable. Should only be created directly in
+        // SignatureSetArray via copy elision
+        SignatureSet(const SignatureSet &source) = delete;
+        SignatureSet(SignatureSet &&source) = default;
+        SignatureSet &operator=(const SignatureSet &source) = delete;
+        SignatureSet &operator=(SignatureSet &&source) = default;
+    };
+
+    class SignatureSetArray : public BlstBase
+    {
+    public:
+        std::vector<SignatureSet> _sets;
+
+        SignatureSetArray(Napi::Env env, const Napi::Value &raw_arg);
+
+        // immovable/non-copyable. should only be created directly as class member
+        SignatureSetArray(const SignatureSetArray &source) = delete;
+        SignatureSetArray(SignatureSetArray &&source) = delete;
+        SignatureSetArray &operator=(const SignatureSetArray &source) = delete;
+        SignatureSetArray &operator=(SignatureSetArray &&source) = delete;
+
+        SignatureSet &operator[](size_t index) { return _sets[index]; }
+
+        size_t Size() { return _sets.size(); }
+    };
+} // namespace (anonymous)
+
 namespace Functions
 {
-    Napi::Value VerifyMultipleAggregateSignatures(const Napi::CallbackInfo &info);
-    Napi::Value VerifyMultipleAggregateSignaturesSync(const Napi::CallbackInfo &info);
-
     void Init(const Napi::Env &env, Napi::Object &exports);
 }
